requirements.txt
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.1
blis==0.7.4
Bottleneck==1.3.2
catalogue==2.0.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==7.1.2
cryptography==3.4.7
cycler==0.10.0
cymem==2.0.5
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
Flask==1.1.2
fpdf==1.7.2
idna==3.2
importlib-metadata==3.10.0
ipykernel==5.3.4
ipython==7.22.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.17.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
murmurhash==1.0.5
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.4.0
numexpr==2.7.3
numpy==1.20.3
olefile==0.46
packaging==21.0
pandas==1.3.0
pandocfilters==1.4.3
parso==0.8.2
pathy==0.6.0
pdf2image==1.16.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pip==21.2.1
preshed==3.0.5
prettytable==2.1.0
prometheus-client==0.11.0
prompt-toolkit==3.0.17
ptyprocess==0.7.0
pycparser==2.20
pydantic==1.8.2
Pygments==2.9.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytesseract==0.3.8
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2021.1
pyzmq==20.0.0
requests==2.26.0
scipy==1.6.2
seaborn==0.11.1
Send2Trash==1.5.0
setuptools==57.4.0
six==1.16.0
smart-open==5.1.0
spacy==3.1.1
spacy-legacy==3.0.8
srsly==2.4.1
terminado==0.9.4
testpath==0.5.0
thinc==8.0.8
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.6
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wheel==0.36.2
zipp==3.5.0
